pyproject.toml
[project]
name = "dinov3-downstream"
version = "0.1.0"
description = "DINOv3 downstream project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "transformers>=4.57.1",
    "torch>=2.4.1",
    "torchvision>=0.18.1",
    "peft>=0.17.1",
    "pytest>=9.0.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
    { index = "pytorch-cpu",     marker = "sys_platform != 'linux' and sys_platform != 'win32'" }
]
torchvision = [
    { index = "pytorch-cu121", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
    { index = "pytorch-cpu",     marker = "sys_platform != 'linux' and sys_platform != 'win32'" }
]
